Warning fixes

+++ Core/HLE/__sceAudio.cpp
@@ -186,7 +186,7 @@ void __AudioUpdate()
 		} else {
 			// This happens quite a lot. There's still something slightly off
 			// about the amount of audio we produce.
-			DEBUG_LOG(HLE, "Audio outbuffer overrun! room = %i / %i", outAudioQueue.room(), outAudioQueue.capacity());
+			DEBUG_LOG(HLE, "Audio outbuffer overrun! room = %i / %i", outAudioQueue.room(), (u32)outAudioQueue.capacity());
 		}
 	}
